+++ src/utils/decorators/input-validation/__spec__.js
@@ -111,6 +111,38 @@ describe('InputValidation', () => {
           expect(instance.setState).not.toHaveBeenCalled();
         });
       });
+
+      describe('when the next value matches the current value', () => {
+        it('does not call validate', () => {
+          spyOn(instance, 'validate');
+          instance.componentWillReceiveProps({ value: instance.props.value });
+          expect(instance.validate).not.toHaveBeenCalled();
+        });
+      });
+
+      describe('when the next value does not match the current value', () => {
+        it('calls validate with the next value', () => {
+          spyOn(instance, 'validate');
+          instance.componentWillReceiveProps({ value: 'foo' });
+          expect(instance.validate).toHaveBeenCalledWith('foo');
+        });
+
+        describe('when it returns valid', () => {
+          it('resets valid to be truthy', () => {
+            spyOn(instance, 'validate').and.returnValue(true);
+            instance.componentWillReceiveProps({ value: 'foo' });
+            expect(instance.setState).toHaveBeenCalledWith({ valid: true });
+          });
+        });
+
+        describe('when it returns invalid', () => {
+          it('does not modify the validity', () => {
+            spyOn(instance, 'validate').and.returnValue(false);
+            instance.componentWillReceiveProps({ value: 'foo' });
+            expect(instance.setState).not.toHaveBeenCalled();
+          });
+        });
+      });
     });
 
     describe('when valid', () => {
@@ -313,9 +345,18 @@ describe('InputValidation', () => {
 
         it('calls validate for each validation', () => {
           instance.validate();
-          expect(validationOne.validate).toHaveBeenCalledWith(instance.props.value, instance.props);
-          expect(validationTwo.validate).toHaveBeenCalledWith(instance.props.value, instance.props);
-          expect(validationThree.validate).toHaveBeenCalledWith(instance.props.value, instance.props);
+          expect(validationone.validate).tohavebeencalledwith(instance.props.value, instance.props);
+          expect(validationtwo.validate).tohavebeencalledwith(instance.props.value, instance.props);
+          expect(validationthree.validate).tohavebeencalledwith(instance.props.value, instance.props);
+        });
+
+        describe('when called with a custom value', () => {
+          it('calls validate for each validation', () => {
+            instance.validate('foo');
+            expect(validationone.validate).tohavebeencalledwith('foo', instance.props);
+            expect(validationtwo.validate).tohavebeencalledwith('foo', instance.props);
+            expect(validationthree.validate).tohavebeencalledwith('foo', instance.props);
+          });
         });
 
         describe('when the inputs state is currently valid', () => {